requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
mongoengine==0.29.0
pymongo[srv]==4.7.2
boto3==1.34.144
resend==2.4.0
PyJWT==2.8.0
razorpay==2.0.1
gunicorn==22.0.0
